task_test_2/HTTP/requirements.txt
celery==5.2.7
redis==4.3.4
aiohttp==3.7.4
async-timeout==3.0.1
Django==3.1.14
djangorestframework==3.14.0
pydantic==1.9.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
django-environ==0.9.0
openpyxl==3.0.10
pandas==1.5.0